Make block a float internally, cast to int
Change EntityFinalStats.block from int to float to allow fractional/block calculations internally. When applying updates, clamp to non-negative and cast to int (int(max(0, entity_stats.block))) before replacing the BlockComponent so the external component remains an integer. This preserves internal precision while keeping stored component values integral.

+++ src/ai_rpg/systems/arbitration_action_system.py
@@ -35,7 +35,7 @@ from ..utils import extract_json_from_code_block
 @final
 class EntityFinalStats(BaseModel):
     hp: float
-    block: int
+    block: float
 
 
 #######################################################################################################################################
@@ -444,7 +444,7 @@ class ArbitrationActionSystem(ReactiveProcessor):
                     f"更新 {entity_name} HP: {old_hp} → {new_hp}/{max_hp}, block: {entity_stats.block}"
                 )
 
-                new_block = max(0, entity_stats.block)
+                new_block = int(max(0, entity_stats.block))
                 entity.replace(BlockComponent, entity_name, new_block)
 
                 self._game.add_human_message(
